chore: move isProduction to helper.tsx
